Перенос функции генерации id
